[#13][front] local game - rename <ce-icon> to <ui-icon>
