+++ front/public/components/icons/icons.js
